refactor: move file processing logic out of main
