+++ packages/cli/src/ezsam/config.py
@@ -7,3 +7,4 @@ DEFAULT_TEXT_THRESHOLD = 0.3
 DEFAULT_NMS_THRESHOLD = 0.8
 # Supports transparency
 OUTPUT_IMAGE_FORMAT = 'png'
+OUTPUT_VIDEO_FORMAT = 'webm'
